refactor: follow the guideline
